+++ src/EspBleClassic.cpp
@@ -1175,7 +1175,7 @@ void EspBleClassicSpp::update()
 
 EspBleClassic::EspBleClassic() :
   spp_(this), hidDevice_(this), hidHost_(this), a2dpSink_(this),
-  a2dpSource_(this), avrcp_(this) {}
+  a2dpSource_(this), avrcp_(this), hfpClient_(this) {}
 
 EspBleClassic::~EspBleClassic()
 {
@@ -1354,6 +1354,7 @@ bool EspBleClassic::begin(const EspBleClassicConfig &config)
 void EspBleClassic::end()
 {
   if (!initialized()) return;
+  hfpClient_.end();
   avrcp_.end();
   a2dpSink_.end();
   a2dpSource_.end();
@@ -1391,6 +1392,7 @@ void EspBleClassic::update()
   a2dpSink_.update();
   a2dpSource_.update();
   avrcp_.update();
+  hfpClient_.update();
 }
 
 bool EspBleClassic::initialized() const
@@ -1426,6 +1428,11 @@ EspBleClassicA2dpSource &EspBleClassic::a2dpSource()
 EspBleClassicAvrcp &EspBleClassic::avrcp()
 {
   return avrcp_;
+}
+
+EspBleClassicHfpClient &EspBleClassic::hfpClient()
+{
+  return hfpClient_;
 }
 
 EspBleError EspBleClassic::lastError() const